backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
openpyxl
python-dotenv
